requirements.txt
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.1
pyjwt==2.3.0
bcrypt==4.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
